requirements.txt
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
tushare>=1.2.0
akshare>=1.10.0
pymysql>=1.0.0
python-dotenv>=1.0.0
openai>=1.0.0
pyyaml>=6.0
